[Breaking] KeyboardManager: Enables custom callbacks in tandem with bind
Currently, if setting custom `.on(event:callback:)` on a keyboard
manager that has been bound to a textInputView, the bound callbacks will
replace the callbacks set by `bind(inputAccessoryView:usingTabBar:)`.
This change allows the two to coexist.

An alternative I considered was turning `KeyboardManager.callbacks` into
an array instead, but that seemed more complex for a consumer of the
API. Open to any alternatives

Since this is a breaking change, apps that were dependant on being able
to override callbacks set by `bind(inputAccessoryView:)` will need to
change their approach. My assumption is that very few apps were using
that workaround, but we should call this out nonetheless